Identify Tables
Added Identify tables for all relevant firemodes on the base firemodes table (adding the rest tomorrow, I hope!) Identify tables more cleanly and easily let the firemodes know which "version" of their effects are being used.
For example, when the Grenade firemode is armed, soda-based models play a can opening sound. Rather than checking if the model is X or Y or Z or ... for every soda model like before, it now checks the model in the Identify table, and plays soda sounds if the table returns a value of 0.
This will both run more efficiently and make it easier to add models to existing "flavors" for existing firemodes.

Updated Stacking to compare firemode names and Identify table values, rather than model and skin. Lacking an identify table means all versions of a firemode are treated as unique for stacking purposes.